+++ CHANGELOG.md
@@ -8,6 +8,7 @@ is `smoothing_penalizer`.
 - Changed `penalizer` keyword arg to `coef_penalizer` in AalenAdditiveFitter.
 - new `ridge_regression` function in `utils.py` to perform linear regression with l2 penalizer terms.
 - Matplotlib is no longer a mandatory dependency. 
+- `.predict(time)` method on univariate fitters can now accept a scalar (and returns a scalar) and an iterable (and returns a numpy array)
 
 #### 0.5.1